@@ -1,4 +1,6 @@
-import { Schema, model } from "mongoose";
+import pkg from "mongoose";
+
+const { Schema, model } = pkg;
 
 const fileSchema = new Schema({
     fileName: String,
